1st Riccardo G. commit
A new view (messages.py) has been created. This should be used to handle everything which deals with the messages (creation, sending, etc.). It has been inserted inside a "/new_message" route which is meant to handle the creation of the messages, but it is still incomplete.

To use the messages.py view, the file views/__init__py. has been modified.

a new template has been created: new_message.html, WHICH PRESENT SOME PROBLEMS DURING THE RENDERING PHASE.
This page makes use of the new form "MessageForm", added in the forms.py file.

the environment for celery has been set up (in the background.py file). It is necessary to add the required tasks in the "tasks" folder, like the "new_message_tasks.py", which contains an example of task.

Contact me for more details