2nd Riccardo G. commit
The main changes are related to the messages.py view.
In particular, the function new_message() has been structured in a way that:
1.  verifies that the user is logged
2. in case of a POST request, it creates an instance of Message(), populate the message, create an instance of the Message_logic class (which has been created to handle all the logic behind the messages (e.g., communicate with the database; perform checks on the received types)) ad the message is passed to the msg_logic object.
3. multiple recipients from the form are singularly taken and, for each of them, a Message_Recipient object is created, populated and managed by the msg_logic.
4. finally, depending on what button is clicked into the form, a different page is rendered

a new route has been created to handle the sending of the message/bottle after having taken a delivery time. However, this route hasn't been completely written yet.

The forms.py file has been updated with the new forms to handle the creation of a new message and its sending (by inserting a delivery time)

Co-Authored-By: GManfry <72792517+GManfry@users.noreply.github.com>