3rd commit Riccardo G. / Gianmarco
The messages.py file has been updated to inlude several checks and the sending of a message.

in the message_logic.py other methods have been created to perform checks and to communicate with the databes (to insert and update messages)

forms.py have been modified. In particular MessageForm now includes also the deliver_time parameter and it has been updated to correctly obtain a list of recipients

Co-Authored-By: GManfry <72792517+GManfry@users.noreply.github.com>