1st commit Riccardo G. Celery
Celery has been used to asynchronously deliver a message and send a notification to the recipient (the notification has been implemented as a print. It will be changed with the sending of an email).
In particular, to create the asynchronous task:
1. the timezone configuration has been set to 'Europe/Rome'
2. a periodic task has been created. This defines the task that will be periodically called and the time to call after which call the task
3. the deliver_message_and_send_notification function has been written to retreive messages from the database that need to be sent, for sending an email  to the recipients and for updating the messages in the database (set the is_delivered attribute to True)

in the message_logic.py file the previously inserted send_notification celery function has been removed. The same has been done in the messages.py file, where the apply_async function was called