Commit Riccardo G.
Added tests for messages.py. To do this, the message logic functions' return values have been changed (not all of them) in order to better perform the tests. Also, in the database some methods have been added to provide better flexibility when invoked from MessageLogic methods.

a script (test.sh) have been created to start the test and deleting the db everytime the test start.

in app.py file a populate_db() function has been created to populate the app with fake data each time the application is run (only if the db doesn't already exist).
Also, a variable test_mode has been added in the create_app() method so to specify if the application is running on test modo or "normal mode"