Riccardo g/giammarco/sahar (#23)
* 1st commit Riccardo G. Celery

Celery has been used to asynchronously deliver a message and send a notification to the recipient (the notification has been implemented as a print. It will be changed with the sending of an email).
In particular, to create the asynchronous task:
1. the timezone configuration has been set to 'Europe/Rome'
2. a periodic task has been created. This defines the task that will be periodically called and the time to call after which call the task
3. the deliver_message_and_send_notification function has been written to retreive messages from the database that need to be sent, for sending an email  to the recipients and for updating the messages in the database (set the is_delivered attribute to True)

in the message_logic.py file the previously inserted send_notification celery function has been removed. The same has been done in the messages.py file, where the apply_async function was called

* commit Riccardo G.

some changes in the test_newmessage.py to understand how to create a new db (but is is incomplete)

added in the background.py file some comments to understand how to send email notifications

* Commit Riccardo G. – celery

Added an emails.py file to start a smtp server (with a secure ssl connection) and send emails. The send_email() method is called from the celery tasks for sending notifications when a new message is sent and when the lottery is carried out (monthly)

to implement the lottery functionality a new column (lottery_ponts) in the User table has been added, which represent the points that each user has (and that can be used for the removal of a sent message before it arrives to the recipient)

* Commit Riccardo G.

Added tests for messages.py. To do this, the message logic functions' return values have been changed (not all of them) in order to better perform the tests. Also, in the database some methods have been added to provide better flexibility when invoked from MessageLogic methods.

a script (test.sh) have been created to start the test and deleting the db everytime the test start.

in app.py file a populate_db() function has been created to populate the app with fake data each time the application is run (only if the db doesn't already exist).
Also, a variable test_mode has been added in the create_app() method so to specify if the application is running on test modo or "normal mode"

* Commit Riccardo G. – test emails

Email tests have been added to check if an email is correctly sent or not