Added calendar functionalities
A new view calendar.py has been added to sho the calendar of received and sent messages.
2 main files have been created to properly handle the creation of the calendar:
- calendar_logic.py
- calendar.html

calendat_logic.py:
it contains a class with two main functions used to retrieve the information that need to be displayed into the calendar (firstname, lastname and email of recipient or sender of the message + deliver_time of the message).

calendar.html:
is a simple html page with some links to the libraries that need to be imported to create a calendar. The main part of this page is the final script tag, which receive from the backend the data to be displayed into the calendar (after a minimal preprocessing phase), create a calendar object and populate it with the required data. Finally, it renders the calendar.

To access the calendar, a new button has been inserted into the homepage