Deleting of a message implemented
Added a route in messages to handle the deletion of a message. It will check if the message id is an existing one, if the user is logged, if the message is sent but not delivered and if the user has enough points to delete the message. If so, it proceeds to the deletion of the message (updating both the db and the eventual folder with images).

A button has been added to the message_detail.html page to delete a message

Also, some comments have been removed from new_messages.html and modify_draft.html