Implemented tests for deletion of messages
messages.py:
part of the content of the /delete_message route has been moved to the MessageLogic class to better perform the tests.

test_newmessage.py:
the file has been renamed test_message.py and now it also contains the tests for the deletion of a message.

test_populate_db.py:
the file has been properly modified to better perform the test of the deletion of a message