Added deletion of messages functionality (#32)
* Deleting of a message implemented

Added a route in messages to handle the deletion of a message. It will check if the message id is an existing one, if the user is logged, if the message is sent but not delivered and if the user has enough points to delete the message. If so, it proceeds to the deletion of the message (updating both the db and the eventual folder with images).

A button has been added to the message_detail.html page to delete a message

Also, some comments have been removed from new_messages.html and modify_draft.html

* Implemented tests for deletion of messages

messages.py:
part of the content of the /delete_message route has been moved to the MessageLogic class to better perform the tests.

test_newmessage.py:
the file has been renamed test_message.py and now it also contains the tests for the deletion of a message.

test_populate_db.py:
the file has been properly modified to better perform the test of the deletion of a message

* Removed useless parts in test_emails.py

* Modified run.sh and test.sh to correctly remove and create the static/attached/ directory once either the application or the test is started

* added some missing tests of delete_message