Temporary branch merge (#41)
* Added functionality to modify personal data

* added tests for modification of personal data

* changed URI from /modify_personal_data in /data

* Implemented functionality to modify the password of an account

forms.py:
added a form for the modification of the password containing 3 password fields (old, new and repetition f new password)

modify_password.html:
page that renders the form for the modification of the password

user_details.html:
contains a new button for the modification of the password

users.py:
added a route to handle the modificatin of the password. GET to retrieve the modification password form. POST for the update of the password.
If some errors occur while modifying the password, the user_logic function will return a specific integer /that is properly hendled depending on the error)

user_logic.py:
added methods for the modification of the password.
-check_form_password --> check if the old password is the same of the one stored in the database, that the old and new password are not the same and that the new password and the repeated new password are different. If no errors occur, value '4' is returned (Which stands for "NO ERRORS")
-modify_password --> set the new password for the current user

* Tested user_logic for the modification of the password

* added rendering tests for the modification of the password

* commit