Extend test messages.py with other tests (including also some tests for insertion of images in a new message) (#34)
* Deleting of a message implemented

Added a route in messages to handle the deletion of a message. It will check if the message id is an existing one, if the user is logged, if the message is sent but not delivered and if the user has enough points to delete the message. If so, it proceeds to the deletion of the message (updating both the db and the eventual folder with images).

A button has been added to the message_detail.html page to delete a message

Also, some comments have been removed from new_messages.html and modify_draft.html

* Implemented tests for deletion of messages

messages.py:
part of the content of the /delete_message route has been moved to the MessageLogic class to better perform the tests.

test_newmessage.py:
the file has been renamed test_message.py and now it also contains the tests for the deletion of a message.

test_populate_db.py:
the file has been properly modified to better perform the test of the deletion of a message

* Removed useless parts in test_emails.py

* Modified run.sh and test.sh to correctly remove and create the static/attached/ directory once either the application or the test is started

* added some missing tests of delete_message

* Added tests for message_logic.py

Added tests for message_logic.py

To better do this, a new class (Image) has been defined to properly check if the message contains an image or not

* Added some tests for the sending of a message with images

Not all the tests have been carried out because of the difficulty to provide an image directly from the post method (and not from a form)

* fix test

* better fix

Co-authored-by: Alessio Matricardi <alessiomatricardi@yahoo.com>