Implementation of the search of a user
users.py: added route /users/search to manage the search of a user. POST request parse the form and return a list of users matching the form fields.

Added a method in list_logic.py for retrieving the list of searched users

Added test_15_search.py for testing the search of a user

Added a form for the search of a user

Added html page for the search of a user

Fixing again the attach image rendering.

Fixing calendar editing.

Added download of spacy en in test.sh

Added some missing methods=['GET']

removed obsolete test in test_09

Fixed Back button graphical rendering.